requirements.txt
python>=3.8
python-docx>=0.8.11
PyPDF2>=3.0.0

